requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
networkx>=2.6.0
torch>=1.9.0
tqdm>=4.62.0
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
jupyter>=1.0.0
plotly>=5.3.0
